+++ src/servers/mod.rs
@@ -365,6 +365,8 @@ pub mod test_setup {
     pub const TEST_PASSWORD_HASH: &str = "$argon2id$v=19$m=4096,t=1,p=1$D/DKFfvJbZOBICD6y/798w$ifr1qDS9aQLyRPT+57ZOKmfUnrju+fbkEpiK6w2ADuo";
     pub const TEST_FULL_NAME: &str = "Test user full name";
 
+    pub const UNSAFE_PASSWORD: &str = "iloveyou1234";
+
     pub const TEST_USER_AGENT: &str = "test_user_agent";
 
     pub const ANON_EMAIL: &str = "anon_user@email.com";
@@ -865,7 +867,7 @@ pub mod test_setup {
 
         /// Get Test users active devices
         pub async fn query_user_active_devices(&self) -> Vec<DeviceQuery> {
-            let query = r"SELECT de.*, ap.api_key_string FROM device de LEFT JOIN api_key ap ON de.api_key_id = ap.api_key_id WHERE de.registered_user_id = $1 AND de.active = true";
+            let query = r"SELECT de.*, ap.api_key_string FROM device de LEFT JOIN api_key ap USING(api_key_id) WHERE de.registered_user_id = $1 AND de.active = true";
             sqlx::query_as::<_, DeviceQuery>(query)
                 .bind(self.model_user.as_ref().unwrap().registered_user_id.get())
                 .fetch_all(&self.postgres)
@@ -874,7 +876,7 @@ pub mod test_setup {
         }
 
         async fn query_anon_user_active_devices(&self) -> Vec<DeviceQuery> {
-            let query = r"SELECT de.*, ap.api_key_string FROM device de LEFT JOIN api_key ap ON de.api_key_id = ap.api_key_id WHERE de.registered_user_id = $1 AND de.active = true";
+            let query = r"SELECT de.*, ap.api_key_string FROM device de LEFT JOIN api_key ap USING(api_key_id) WHERE de.registered_user_id = $1 AND de.active = true";
             sqlx::query_as::<_, DeviceQuery>(query)
                 .bind(self.anon_user.as_ref().unwrap().registered_user_id.get())
                 .fetch_all(&self.postgres)
@@ -955,6 +957,13 @@ pub mod test_setup {
             .unwrap()
         }
 
+        // device de
+        // ON
+        // co.device_id = de.device_id
+        // LEFT JOIN
+        // ip_address ipa
+        // ON
+        // co.ip_id = ipa.ip_id
         /// will sleep before query!
         pub async fn get_connections(
             &self,
@@ -967,18 +976,13 @@ pub mod test_setup {
 				co.connection_id, co.timestamp_online::TEXT, co.timestamp_offline::TEXT
 			FROM
 				connection co
-			LEFT JOIN
-				device de
-			ON
-				co.device_id = de.device_id
+			LEFT JOIN ip_address ipa USING(ip_id)
+			LEFT JOIN device de USING(device_id)
 			LEFT JOIN
 				device_name dn
 			ON
 				de.device_name_id = dn.device_name_id
-			LEFT JOIN
-				ip_address ipa
-			ON
-				co.ip_id = ipa.ip_id
+		
 			WHERE
 				de.registered_user_id = $1
 			AND