+++ CHANGELOG.md
@@ -1,2 +1,22 @@
+# <a href='https://github.com/mrjackwills/staticpi_backend/releases/tag/v0.0.1'>v0.0.1</a>
+### 2023-03-06
+
+### Chores
++ dependencies updated, [d7a5d5e7](https://github.com/mrjackwills/staticpi_backend/commit/d7a5d5e73b781636f8d75f80ac50957e8eb3ae84),, [e3490c8d](https://github.com/mrjackwills/staticpi_backend/commit/e3490c8d6cdc1818594581400cac6da3e29758ac),
++ devcontainer updated, [e907057f](https://github.com/mrjackwills/staticpi_backend/commit/e907057f1321839f8df1ecc2d57a9f2d9c20fefc),
+
+### Docs
++ readme updated, [709477af](https://github.com/mrjackwills/staticpi_backend/commit/709477af645d2851598d1b49b6803eda8fa906e2),
+
 ### Features
-+ init commit
++ argon2 param builder into lazy static, [eeba93ea](https://github.com/mrjackwills/staticpi_backend/commit/eeba93eac397ab75d71fa26c4de3d2322502accc),
+
+### Fixes
++ incoming json downcast error, [7699db04](https://github.com/mrjackwills/staticpi_backend/commit/7699db04cb122e785bf95c191a3187e52edfd97c),
++ devcontainer updated, [31acce16](https://github.com/mrjackwills/staticpi_backend/commit/31acce167d0efe49857f261302b5eb63da4a89e1),
+
+### Refactors
++ postgreSQL use `USING(x)` where appropiate, [1018a280](https://github.com/mrjackwills/staticpi_backend/commit/1018a280bca4af88ab31713cb8f99c8dedca463c),
+
+### Tests
++ use UNSAFE_PASSWORD const, [3a061a07](https://github.com/mrjackwills/staticpi_backend/commit/3a061a07a7ccdf980fa53c9296a519d78f407192),